backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-multipart==0.0.6
azure-communication-email==1.0.0
azure-identity==1.15.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
